Implement two-copy system for bulk set and insert buffer operations

+++ src/main/scala/org/kneelawk/learningopengl/buffers/GLArrayBufferSetOperation.scala
@@ -7,7 +7,7 @@ import org.lwjgl.system.MemoryUtil
 /**
  * Object representing a set operation.
  */
-case class GLArrayBufferSetOperation(offset: Long, bufLen: Long, bufData: Long) {
+case class GLArrayBufferSetOperation(offset: Int, bufLen: Int, bufData: Long) {
 
   if (offset < 0)
     throw new IllegalArgumentException("The offset cannot be negative")
@@ -18,36 +18,36 @@ case class GLArrayBufferSetOperation(offset: Long, bufLen: Long, bufData: Long) 
   /**
    * ByteBuffer version of the constructor.
    */
-  def this(offset: Long, buf: ByteBuffer) =
+  def this(offset: Int, buf: ByteBuffer) =
     this(offset, buf.remaining(), MemoryUtil.memAddress(buf))
 
   /**
    * ShortBuffer version of the constructor.
    */
-  def this(offset: Long, buf: ShortBuffer) =
+  def this(offset: Int, buf: ShortBuffer) =
     this(offset, buf.remaining() << 1, MemoryUtil.memAddress(buf))
 
   /**
    * IntBuffer version of the constructor.
    */
-  def this(offset: Long, buf: IntBuffer) =
+  def this(offset: Int, buf: IntBuffer) =
     this(offset, buf.remaining() << 2, MemoryUtil.memAddress(buf))
 
   /**
    * LongBuffer version of the constructor.
    */
-  def this(offset: Long, buf: LongBuffer) =
+  def this(offset: Int, buf: LongBuffer) =
     this(offset, buf.remaining() << 3, MemoryUtil.memAddress(buf))
 
   /**
    * FloatBuffer version of the constructor.
    */
-  def this(offset: Long, buf: FloatBuffer) =
+  def this(offset: Int, buf: FloatBuffer) =
     this(offset, buf.remaining() << 2, MemoryUtil.memAddress(buf))
 
   /**
    * DoubleBuffer version of the constructor.
    */
-  def this(offset: Long, buf: DoubleBuffer) =
+  def this(offset: Int, buf: DoubleBuffer) =
     this(offset, buf.remaining() << 3, MemoryUtil.memAddress(buf))
 }
